+++ docs-site/.boltrc.js
@@ -11,6 +11,7 @@ const config = {
   openServerAtStart: true,
   // Environmental variable / preset to use
   env: 'pwa',
+  esModules: true,
   srcDir: './src/pages',
   buildDir: '../www/build',
   dataDir: '../www/build/data',
@@ -70,7 +71,10 @@ const config = {
 
   components: {
     global: [
+      // helper components that are only used internally
+      '@bolt/shadow-toggle',
       '@bolt/theme-switcher',
+
       '@bolt/components-radio-switch',
       '@bolt/components-carousel',
       '@bolt/global',
@@ -95,6 +99,7 @@ const config = {
       '@bolt/components-button',
       '@bolt/components-button-group',
       '@bolt/components-card',
+      '@bolt/components-card-replacement',
       '@bolt/components-chip',
       '@bolt/components-chip-list',
       '@bolt/components-code-snippet',
@@ -159,7 +164,10 @@ const config = {
       to: path.join(__dirname, '../www/build'),
     },
     {
-      from: path.join(path.dirname(require.resolve(`@bolt/components-typeahead`)),'__demos__/typeahead.data.json'),
+      from: path.join(
+        path.dirname(require.resolve(`@bolt/components-typeahead`)),
+        '__demos__/typeahead.data.json',
+      ),
       to: path.join(__dirname, '../www/build/data'),
     },
     {
@@ -174,7 +182,7 @@ const config = {
     },
     {
       from: `${path.dirname(
-        resolve.sync('@bolt/global/package.json')
+        resolve.sync('@bolt/global/package.json'),
       )}/favicons/bolt`,
       to: path.join(__dirname, '../www/'),
       flatten: true,
@@ -183,7 +191,7 @@ const config = {
   alterTwigEnv: [
     {
       file: `${path.dirname(
-        resolve.sync('@bolt/twig-renderer/package.json')
+        resolve.sync('@bolt/twig-renderer/package.json'),
       )}/SetupTwigRenderer.php`,
       functions: ['addBoltCoreExtensions', 'addBoltExtraExtensions'],
     },