+++ packages/components/bolt-icons/src/icons/reply.js
@@ -1,4 +1,5 @@
-import { h } from '@bolt/core/renderers';
+import * as Icons from '@bolt/components-icon/registry';
+import { h } from '@bolt/core-v3.x/renderers';
 
 export const Reply = ({ bgColor, fgColor, size, ...otherProps }) => {
   return (
@@ -7,3 +8,5 @@ export const Reply = ({ bgColor, fgColor, size, ...otherProps }) => {
     </svg>
   );
 };
+
+Icons.set('reply', Reply);