+++ packages/components/bolt-nav-priority/index.js
@@ -1,4 +1,4 @@
-import { polyfillLoader } from '@bolt/core/polyfills';
+import { polyfillLoader } from '@bolt/core-v3.x/polyfills';
 
 polyfillLoader.then(() => {
   import('./nav-priority.js');