+++ packages/experimental/micro-journeys/src/utils/is-chrome.js
@@ -1,0 +1,28 @@
+/**
+ * Determine if this browser is Google Chrome or not.
+ * https://stackoverflow.com/a/13348618 for why this logic is necessary.
+ *
+ * @return {boolean}: true if Chrome, false if not.
+ */
+const isChrome = () => {
+  const isChromium = window.chrome;
+  const winNav = window.navigator;
+  const vendorName = winNav.vendor;
+  const isOpera = typeof window.opr !== 'undefined';
+  const isIEedge = winNav.userAgent.indexOf('Edge') > -1;
+  const isIOSChrome = winNav.userAgent.match('CriOS');
+  if (isIOSChrome) {
+    return true;
+  } else if (
+    isChromium !== null &&
+    typeof isChromium !== 'undefined' &&
+    vendorName === 'Google Inc.' &&
+    isOpera === false &&
+    isIEedge === false
+  ) {
+    return true;
+  }
+  return false;
+};
+
+export default isChrome;